Refactored ecommerce installer into a service.
