+++ presto.install
@@ -9,7 +9,7 @@ use Drupal\Core\Language\LanguageInterface;
 use Drupal\menu_link_content\Entity\MenuLinkContent;
 use Drupal\node\Entity\Node;
 use Drupal\presto\Form\PrestoConfigureForm;
-use Drupal\presto\Installer\Ecommerce\Installer as EcommerceInstaller;
+use Drupal\presto\Installer\Ecommerce\EcommerceInstaller;
 use Drupal\shortcut\Entity\Shortcut;
 use Drupal\user\Entity\User;
 use Drupal\user\RoleInterface;
@@ -136,7 +136,9 @@ function presto_apply_configuration(array &$install_state) {
     'operations' => [],
   ];
 
-  $ecommerceInstaller = EcommerceInstaller::create((array)$install_state);
+  /** @var Drupal\presto\Installer\Ecommerce\EcommerceInstaller $ecommerceInstaller */
+  $ecommerceInstaller = Drupal::service('presto.installer.ecommerce');
+  $ecommerceInstaller->setInstallState((array) $install_state);
   // Install modules and create demo content if eCommerce is enabled.
   $batch['operations'] = array_merge(
     $batch['operations'],
