+++ src/Installer/InstallerException.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace Drupal\presto\Installer\Ecommerce;
+namespace Drupal\presto\Installer;
 
 use Exception;
 
 /**
  * Thrown if the installer fails.
  *
- * @package Drupal\presto\Installer\Ecommerce
+ * @package Drupal\presto\Installer
  */
 class InstallerException extends Exception {
 }
